Move back AttributeValue and Status from data to the main package.
