+++ api/src/main/java/openconsensus/trace/SpanBuilder.java
@@ -18,7 +18,6 @@ package openconsensus.trace;
 
 import java.util.List;
 import java.util.concurrent.Callable;
-import openconsensus.trace.data.Status;
 
 /**
  * {@link SpanBuilder} is used to construct {@link Span} instances which define arbitrary scopes of
