+++ api/src/main/java/openconsensus/trace/Tracer.java
@@ -21,7 +21,6 @@ import java.util.concurrent.Callable;
 import javax.annotation.Nullable;
 import openconsensus.context.Scope;
 import openconsensus.trace.data.SpanData;
-import openconsensus.trace.data.Status;
 import openconsensus.trace.propagation.BinaryFormat;
 import openconsensus.trace.propagation.TextFormat;
